pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fmi-excel-guard"
version = "0.1.0"
description = "Streamlit app for strict FMI Excel upload validation."
requires-python = ">=3.11"
dependencies = [
  "beautifulsoup4>=4.12.3",
  "openai>=1.75.0",
  "openpyxl>=3.1.5",
  "pandas>=2.2.3",
  "python-docx>=1.1.2",
  "streamlit>=1.44.1",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
